touch open
