+++ predicators/ground_truth_models/touch_open/__init__.py
@@ -1,0 +1,5 @@
+"""Ground-truth models for touch open environment and variants."""
+
+from .nsrts import TouchOpenGroundTruthNSRTFactory
+
+__all__ = ["TouchOpenGroundTruthNSRTFactory"]
